+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog for ip2proxy
 ======================
 
+v1.0.1
+------
+
+* Added support for PX9-PX11 BIN database.
+* Added get_package_version and get_database_version funtions to replace the header function.
+* Modify the example in README.md.
+
 v1.0.0
 ------